Added heavy refactoring
