+++ requirements-dev.txt
@@ -1,8 +1,9 @@
 deepdiff>=6.7.1
 mypy>=1.10.0
 pre_commit>=3.5.0
-pytest>=8.0.2
+pytest>=8.2.0
 pytest_cov>=4.1.0
+pytest_mock>=3.14.0
 requests>=2.31.0
 ruff>=0.4.2
 tomli>=2.0.1
